+++ packages/server/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @teloce/server
 
+## 1.1.0
+
+### Minor Changes
+
+- fixes and bugs fix
+
+### Patch Changes
+
+- Updated dependencies
+  - @teloce/shared@1.1.0
+
+## 1.0.0
+
+### Major Changes
+
+- a fixes that was included
+
+### Patch Changes
+
+- Updated dependencies
+  - @teloce/shared@1.0.0
+
 ## 0.2.0
 
 ### Minor Changes