+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @teloce/shared
 
+## 1.1.0
+
+### Minor Changes
+
+- fixes and bugs fix
+
+## 1.0.0
+
+### Major Changes
+
+- a fixes that was included
+
 ## 0.2.0
 
 ### Minor Changes