+++ packages/veloce/CHANGELOG.md
@@ -1,5 +1,17 @@
 # teloce
 
+## 1.1.0
+
+### Minor Changes
+
+- fixes and bugs fix
+
+## 1.0.0
+
+### Major Changes
+
+- a fixes that was included
+
 ## 0.3.0
 
 ### Minor Changes